requirements.txt
attrs==19.3.0
boto==2.49.0
boto3==1.9.134
botocore==1.12.188
certifi==2019.6.16
cffi==1.14.0
chardet==3.0.4
ciso8601==2.1.3
cryptography==2.8
docutils==0.15.2
ephem==3.7.7.1
idna==2.8
importlib-metadata==1.7.0
influxdb==5.3.0
influxdb-client==1.8.0
iniconfig==1.0.0
jmespath==0.9.4
jplephem==2.14
more-itertools==8.4.0
msgpack==0.6.1
numpy==1.19.1
packaging==20.4
pluggy==0.13.1
psycopg2-binary==2.8.4
py==1.9.0
pycparser==2.20
PyJWT==1.6.4
pyparsing==2.4.7
pytest==6.0.0
python-dateutil==2.8.0
pytz==2020.1
requests==2.22.0
Rx==3.1.0
s3transfer==0.2.1
sgp4==2.12
six==1.12.0
skyfield==1.24
toml==0.10.1
tqdm==4.32.2
urllib3==1.25.3
zipp==3.1.0
